requirements.txt
colorlog
matplotlib
sklearn
torchvision

certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
dill==0.3.1.1
Flask==1.1.2
geographiclib==1.50
geopy==1.21.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.1
pycparser==2.20
pycurl==7.43.0.6
pyroute2==0.5.14
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
RestrictedPython==5.0
six==1.15.0
sly==0.3
SQLAlchemy==1.3.19
torch==1.7.0
tqdm
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.0
